% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase-sparql.R
\name{add_statement}
\alias{add_statement}
\title{Add a statement to an item}
\usage{
add_statement(
  data = NULL,
  available_statements = statements,
  new_statement = NULL,
  verbose = TRUE,
  qid_from_row = FALSE,
  col_for_row_content = NULL
)
}
\arguments{
\item{data}{dataframe to add the new column}

\item{available_statements}{dataframe with all statements that can be added with this method. It assumes there is a dataframe that has at least three columns: (1) \code{statement}, (2) \code{pid}, (3) \code{qid}.}

\item{new_statement}{string of new statement. Must exist in \code{available_statements$statement}. If statement is \code{coordinates}, then a column \code{longitude} and a column \code{latitude} is expected.}

\item{verbose}{show in terminal what was added}

\item{qid_from_row}{boolean; default \code{FALSE} - then QID is taken from dataframe \code{statements}, if \code{TRUE} id QID value should be taken from another row}

\item{col_for_row_content}{string; name of column in dataframe \code{data} that contains the QID values}
}
\description{
This function helps to add certain statements. A statement consists of a PID-QID combination that is added to a dataframe. The PID will be the column name, QID the row content.
It assumes there is a dataframe that has at least three columns: (1) \code{statements}, (2) \code{pid}, (3) \code{qid}.
}
\examples{
statements <- data.frame(statement = c("my_statement"), pid = c("P2"), qid = c("Q1"))
data <- data.frame(item = "my item")
data \%>\% add_statement(available_statements = statements, new_statement = "my_statement")
}
