% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-resolution.R
\name{calc_similarity}
\alias{calc_similarity}
\title{Calculate Similarities}
\usage{
calc_similarity(data, method, min_sim)
}
\arguments{
\item{data}{data as a document frequency matrix \code{dfm()} with a set \code{doc_id}}

\item{method}{character; the method identifying the similarity or distance measure to be used, see \code{?quanteda::textstat_simil}}

\item{min_sim}{numeric; a threshold for the similarity values below which similarity values will not be returned; 0.75-0.8 seems reasonable}
}
\value{
dataframe containing the two id's and the similarity value
}
\description{
This functions compute similarities between entries in a document frequency matrix \code{dfm()} and return a dataframe with distinct id combinations.
It heavily relies on the quanteda package
}
