% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-parsing.R
\name{call_lobid_api}
\alias{call_lobid_api}
\title{Call lobid API}
\usage{
call_lobid_api(query, parameter = NULL, verbose = TRUE, as_list = FALSE)
}
\arguments{
\item{query}{string}

\item{parameter}{string, e.g. isbn}

\item{verbose}{if TRUE, then url printed to console}

\item{as_list}{default: \code{FALSE}. If \code{TRUE} then json is converted to a R list, otherwise just url}
}
\value{
list
}
\description{
see: http://lobid.org/resources/api
}
\examples{
\dontrun{
call_lobid_api(3894090685, query = ., parameter = "isbn")
}
}
