% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-text.R
\name{clean_string}
\alias{clean_string}
\title{Clean string}
\usage{
clean_string(data, col)
}
\arguments{
\item{data}{dataframe}

\item{col}{string; column name to clean}
}
\description{
\itemize{
\item remove leading and trailing whitespace
\item change all characters to their lowercase representation
\item remove all punctuation and numbers
\item transforming all special characters (ä, ß, ...) to 1) de-ASCII and 2) Latin-ASCII (see \code{stri_trans_list()})
}
}
\examples{
df <- data.frame(name = "Fritz Müller-Scherz 2")
clean_string(df, "name")
}
