% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_to_distinct_rows}
\alias{collapse_to_distinct_rows}
\title{Collapse ID cols to distinct rows}
\usage{
collapse_to_distinct_rows(...)
}
\arguments{
\item{...}{string or numeric vectors, typically id columns}
}
\description{
Imagine you have a dataframe with two id columns that contain two rows that actually are just one combination.
This function distincts them within a \code{purrr} workflow, see example.
}
\examples{
\dontrun{
df <- data.frame(id_1 = c(1,2), id_2 = c(2,1))
df \%>\%
  dplyr::mutate(id_1_id_2 = purrr::map2_chr(id_1, id_2, collapse_to_distinct_rows)) \%>\%
  dplyr::distinct(id_1_id_2, .keep_all = TRUE) \%>\%
  dplyr::select(-id_1_id_2)
  }
}
\seealso{
Copied from \url{https://stackoverflow.com/a/62381665/2646974}
}
