% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to Databases}
\usage{
connect_db(
  credential_name = "db_local",
  package = "DBI",
  config_path = "~/config.yml",
  default.file.windows = "~/.my.cnf"
)
}
\arguments{
\item{credential_name}{Credential identification in the configuration file,
example \code{datascience}, \code{factfield_data} or \code{pubmed}}

\item{package}{The package to be used: DBI or dbx}

\item{config_path}{Path to the credential file,
example: \code{"~/dbconfig.yml"}}

\item{default.file.windows}{Path to the configuration file in windows
example: \code{"~/.my.cnf"}}
}
\description{
Used to connect to the databases using a configuration file - for unix and windows
The credentials are stored in an external file in the home directory \code{~}.
On Unix: Edit the credentials with \code{file.edit("~/dbconfig.yml")}.
On Windows: Edit the credentials with \code{file.edit("~/.my.cnf")}.
}
\details{
\strong{What inside .yml file?}
\describe{
\item{default:}{}
\item{factfield_data:}{}
\item{host: 'db_main.factfield.de'}{}
\item{user: '........'}{}
\item{password: '......'}{}
\item{port: 13306}{}
\item{dbname: 'db_main'}{}
}
\strong{What inside .cnf file?}

Something like that: \url{https://gist.github.com/rhtyd/d59078be4dc88123104e}
}
\examples{
connect_db() # this connect to factfield_data db_main auto using the credential from the yml data.
}
