% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-columns-to-numeric.R
\name{convert_columns_to_numeric}
\alias{convert_columns_to_numeric}
\title{Converts specific columns to numeric.}
\usage{
convert_columns_to_numeric(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
At first, all imported columns are characters/strings. But of course, most data in the columns
are numeric values. This function makes an educated guess to find all columns that contain
numeric values.
}
\details{
This is done in several steps:
\itemize{
\item make all values to label missing values NA
\item set decimal marks to .
\item guess which colums should be numeric by setting as.numeric if there are no A-z characters in the column
}
}
