% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-text.R
\name{create_fingerprint}
\alias{create_fingerprint}
\title{Create Fingerprint of a string}
\usage{
create_fingerprint(string, tokens = "word", n = NULL)
}
\arguments{
\item{string}{input string}

\item{tokens}{how to generate tokens? \code{word} for whitespace-separated tokens, \code{ngram} for ngrams/shingles}

\item{n}{The number of characters in each shingle. If \code{token = "ngram"} a \code{n} must be provided}
}
\value{
character string
}
\description{
This function creates a fingerprint of a string. This can be used for de-duplication or calculation of string similarity or string distance. It is bases on normalised tokens and implements Open Refine's clustering algorithm, precisly the Fingerprint Key Collision
See \href{Open Refine Clustering Documentation}{https://github.com/OpenRefine/OpenRefine/wiki/Clustering-In-Depth}
}
\examples{
create_fingerprint("Max Spohr Verlag", token = "word")
create_fingerprint("Max Spohr Verlag", token = "ngram", n = 2)
}
