% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-text.R
\name{extract_id}
\alias{extract_id}
\title{Extract QID from string}
\usage{
extract_id(string, pattern = "Q[0-9]+")
}
\arguments{
\item{string}{string to extract from}

\item{pattern}{default for QID}
}
\value{
string
}
\description{
Extract QID from string
}
