% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-parsing.R
\name{extract_text}
\alias{extract_text}
\title{extract text from xml nodes}
\usage{
extract_text(node, class)
}
\arguments{
\item{node}{node}

\item{class}{css class to parse}
}
\description{
When scraping data with \code{rvest} one will typically extract text from nodes.
This functions helps doing so. It makes sure, that multiple \code{p} are parsed into one object.
This is important, when parsing different-length items into one dataframe.
}
\examples{
\dontrun{
"https://forummuenchen.org/lgbtiq-chronik/" \%>\%
  read_html() \%>\%
  html_nodes(".timeline-item") \%>\%
  extract_text("p")
}
}
