% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-parsing.R
\name{get_field_values}
\alias{get_field_values}
\title{Get value of a special field}
\usage{
get_field_values(input, input_type = "url", jq_syntax)
}
\arguments{
\item{input}{the json url}

\item{input_type}{\code{url} when string, \code{response} when fetched response (e.g. when using the same response for multiple queries)}

\item{jq_syntax}{jq filter (test here https://jqplay.org/)}
}
\description{
This function fetches the value of fields in a nested json, no matter on which level.
Based on the very popular js, JSON command line processor https://stedolan.github.io/jq/
}
\examples{
\dontrun{
"https://lobid.org/resources/search?q=isbn:3596237785&format=json" \%>\%
get_field_values("gndIdentifier")
}
}
