% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kabrutils-package.R
\docType{package}
\name{kabrutils-package}
\alias{kabrutils}
\alias{kabrutils-package}
\title{kabrutils: Utility functions for my projects}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Katharina Brunner \email{first.last@example.com}

}
\keyword{internal}
