% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikibase-sparql.R
\name{long_for_quickstatements}
\alias{long_for_quickstatements}
\title{Longform for import}
\usage{
long_for_quickstatements(data, start_at = 2)
}
\arguments{
\item{data}{dataset}

\item{start_at}{column to start longform; input to \code{tidyr::pivot_longer(col)}}
}
\description{
in form: item - property - value
}
