% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-text.R
\name{remove_lang}
\alias{remove_lang}
\title{Remove lang desc
When getting data from a SPARQL endpoint each string has a \verb{@lang} element. This function removes it.}
\usage{
remove_lang(string, pattern)
}
\arguments{
\item{string}{string to remove from}

\item{pattern}{default for QID}
}
\value{
string
}
\description{
Remove lang desc
When getting data from a SPARQL endpoint each string has a \verb{@lang} element. This function removes it.
}
