% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-in-excel.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Open a dataframe in an Excel file}
\usage{
show_in_excel(data)
}
\arguments{
\item{data}{dataframe to show}
}
\description{
Pipe a dataframe to an xlsx file and save it maybe.
}
\examples{
\dontrun{
mtcars \%>\% filter(am == 1) \%>\% show_in_excel()
}
}
