% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-text.R
\name{split_human_name}
\alias{split_human_name}
\title{Split a human full name in its parts}
\usage{
split_human_name(data, col = "name")
}
\arguments{
\item{data}{dataframe containing a column with a human name}

\item{col}{string; name of the column containing the name, defaults to \code{name}}
}
\value{
dataframe with 4 new colums: lastname, givenname, middlename, initial
}
\description{
Rules:
}
\details{
\itemize{
\item lastname = last word
\item givenname = first word
\item middlename = everything not first word or last word
\item initial = first character
}

Beware that there will be a warning \verb{argument is not an atomic vector; coercing }
It works nonetheless.

In general it might \strong{not} be a good idea to try to split a human name, because it's nearly impossible to do that correctly.
Mainly works only for Kartoffel names.
}
\examples{
df <- data.frame(name = c("Rita Mae Brown", "vorname df dsafasf", "sdf asdfd"))
split_human_name(df)
}
\seealso{
\url{https://shinesolutions.com/2018/01/08/falsehoods-programmers-believe-about-names-with-examples/}
}
