% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-parsing.R
\name{transform_json_to_dataframe}
\alias{transform_json_to_dataframe}
\title{Transform json to a tidy dataframe
Uses \code{jsonlite::fromJSON} together with some \code{map}, \code{enframe} and \code{unnest_wider} to create a tidy tibble/dataframe.}
\usage{
transform_json_to_dataframe(json, unnest_type = "long", keep_name = TRUE)
}
\arguments{
\item{json}{json string}

\item{unnest_type}{should the dataframe be wide or long?}

\item{keep_name}{default: \code{TRUE}; \code{tibble::enframe()} names resulting cols, decide if \code{name} column is needed}
}
\value{
dataframe/tibble
}
\description{
Transform json to a tidy dataframe
Uses \code{jsonlite::fromJSON} together with some \code{map}, \code{enframe} and \code{unnest_wider} to create a tidy tibble/dataframe.
}
\examples{
\dontrun{
res <- httr::GET("https://lobid.org/gnd/search?q=preferredName:Max\%20Spohr&format=json")
res <- httr::content(res, as = "text")
df_from_json <- transform_json_to_tidy(res, unnest_type = "wide")
}
}
