% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write a dataframe to an xlsx file}
\usage{
write_xlsx(data, filename, sheet)
}
\arguments{
\item{data}{dataframe to export}

\item{filename}{name of excel file}

\item{sheet}{sheetname}
}
\description{
adds filter and auto-width
}
\details{
beware: deletes existing file with the same name
}
\examples{
\dontrun{
write_xlsx(my_data, "name.xlsx", "sheetname")
}

}
